cast mask_integral in random loc. to proper integral dtype

+++ gunpowder/nodes/random_location.py
@@ -123,7 +123,8 @@ class RandomLocation(BatchFilter):
             logger.debug("chose %s as integral array dtype", mask_integral_dtype)
 
             self.mask_integral = np.array(mask_data > 0, dtype=mask_integral_dtype)
-            self.mask_integral = integral_image(self.mask_integral)
+            self.mask_integral = integral_image(self.mask_integral,
+                                                dtype=mask_integral_dtype)
 
         if self.ensure_nonempty:
